Revert "Navigation Implementation and Organizational Overhaul"
