+++ .gitignore
@@ -8,7 +8,6 @@ npm-debug.*
 *.mobileprovision
 *.orig.*
 web-build/
-yarn-fix.bat
 
 # macOS
 .DS_Store
