--- config/settings.js
@@ -1,2 +1,0 @@
-import React from 'react';
-// put the url to the DB server here
