+++ package.json
@@ -8,20 +8,12 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@react-native-community/masked-view": "0.1.10",
-    "@react-navigation/bottom-tabs": "^5.11.11",
-    "@react-navigation/native": "^5.9.4",
-    "@react-navigation/stack": "^5.14.5",
     "expo": "~41.0.1",
     "expo-status-bar": "~1.0.4",
-    "node-fetch": "^2.6.1",
-    "react": "^17.0.2",
+    "react": "16.13.1",
     "react-dom": "16.13.1",
-    "react-native": "0.63",
-    "react-native-gesture-handler": "^1.10.3",
-    "react-native-safe-area-context": "^3.2.0",
-    "react-native-screens": "~3.0.0",
-    "react-native-web": "^0.16.5",
+    "react-native": "^0.64.2",
+    "react-native-web": "~0.13.12",
     "react-navigation": "^4.4.4"
   },
   "devDependencies": {
