--- screens/CreateReminder.js
@@ -1,14 +1,0 @@
-//CreateReminder screen
-import React from 'react';
-import { Text } from 'react-native';
-import ScreenContainer from '../components/ScreenContainer';
-
-const CreateReminder = ({ navigation }) => {
-  return (
-    <ScreenContainer>
-      <Text>Create a Reminder</Text>
-    </ScreenContainer>
-  )
-}
-
-export default CreateReminder;
