--- screens/Splash.js
@@ -1,14 +1,0 @@
-//Splash screen
-import React from 'react';
-import { Text } from 'react-native';
-import ScreenContainer from "../components/ScreenContainer";
-
-const Splash = () => {
-  return (
-    <ScreenContainer>
-      <Text>Loading</Text>
-    </ScreenContainer>
-  )
-}
-
-export default Splash;
