--- screens/UserDetails.js
@@ -1,14 +1,0 @@
-//UserDetails screen
-import React from 'react';
-import { Text } from 'react-native';
-import ScreenContainer from "../components/ScreenContainer";
-
-const UserDetails = () => {
-  return (
-    <ScreenContainer>
-      <Text>Account information and phone number connects go here</Text>
-    </ScreenContainer>
-  )
-}
-
-export default UserDetails;
