--- screens/index.js
@@ -1,7 +1,0 @@
-// index for screens
-// Make sure you update this file if you move screens around
-export { default as Main } from './Main';
-export { default as CreateReminder } from './CreateReminder';
-export { default as ReminderList } from './ReminderList';
-export { default as UserDetails } from './UserDetails';
-export { default as Splash } from './Splash';
